fix: 멤버 히스토리 벌점 배지 오표시 수정 (#594)
- totalPenalty/totalReward를 서버 누적값(member.penalty)이 아닌
  penaltyHistory 아이템 직접 집계로 변경 → 히스토리가 없으면 배지 미표시
- OperatorMemberDetailSheetViewModel 분리: 상태·액션 로직 위임
- PointGrantFormSheet 별도 파일로 분리, ViewModel 공유로 상태 동기화
- historyAreaHeight(for:) 헬퍼 추출로 동적 높이 계산 중복 제거
- ChallengerMemberDetailSheetView, OperatorMemberDetailSheetView
  Components/Member → Views/Operation 으로 위치 이동
